Add GitHub Actions CI to Build/Release Spec
Add a GitHub Actions CI pipeline that builds the PDF version of the spec
from the source Markdown.  (This uses a combination of a Pandoc docker
image followed by a LaTeX docker image to temporarily work around a bug
in the Pandoc docker image [1].)  The results of the build are uploaded
as a GitHub Actions artifact with an expiration of one day.  This is
done to enable users to check the output of what they built (or to debug
problems where the local installation of pandoc + LaTeX may differ from
what runs in CI).  Finally, if the commit is tagged then a new release
is created where the built PDF spec is added as an additional file.

Fixes #2.

[1]: https://github.com/pandoc/dockerfiles/issues/167

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>