Add script to generate contributors list
Add a shell script that will query GitHub and return a Markdown list of
all users that have contributed to the FIRRTL Specification.  This is
added to provide a simple way to generate an up-to-date list of all
contributors.

This does not automate the updating of the specification to include
these contributors.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>