Correct issues with module description (#14)
The EBNF for a `Module` shows that it contains a list of
statements rather than a single statement. This commit corrects
the description in the Module section to reflect this fact. Prior
to this commit, it implies that a Module only contains a single
statement.

There was a word missing at this location in the doc. I
think "direction" is the right term to describe `input`/`output`
from the grammar.