Add an optional enable signal to registers.
Representation of enable as a mux feeding a register is insufficient as specific tools care about the form of the output for reasoning about enable.  Specifically, they need to be output as if blocks, not as muxes.  However, if and mux have different x behavior in verilog, making it not safe to generally replace a mux with an if.  Further, there is a desire to distingish between a mux before a register and a register with an enable port, making automatic conversion not possible.  Explicitly encoding the enable signal solves both these problems and allows direct expression of intent.