Use git-tag to compute spec version
Change the way that the version is computed to be derived from the
latest git tag that looks like 'v*.*.*'.  This _should_ compose
correctly with the GitHub Action to build/publish the spec once a new
tag is created.  This also has the benefit of never having to manually
update the spec version.  Instead, only a new tag has to be created.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>