Move Revision History to YAML, use git tag version
Change the Revision History section to be encoded using YAML instead of
Markdown.  Update the Pandoc LaTeX template to generate the Revision
History section from this.  The YAML includes two main keys: (1)
thisVersion and (2) oldVersions.  The former is just an array of strings
and will use the latest git tag.  The latter is an array of objects
encoding the version and any changes that should be included.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>