[nfc] Fix, Simplify GitHub CI for auto-versioning
Simplify GitHub CI to not use a bunch of docker containers.  Install
dependencies directly from Ubuntu repos.  This enables the CI to use the
same Makefile that developers are working with and prevent issues where
the Makefile and CI get out of sync.

Fix a bug where the Makefile (or a fixed CI flow) requires the `git
describe` to work.  This does NOT work with a shallow clone.  Bump the
GitHub checkout actions to its latest version (v3) and tell it to fetch
everything.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>