[patch] Add Explicit Info About Aggregate Types
Add clarifications to the spec that FIRRTL has both ground types and
aggregate types.  The latter term "aggregate types" is used, but not
defined.  Also, restructure the spec to include an explicit "aggregate
types" section that is at the same level as "ground types"  Move vector
and bundle types under this new "aggregate types" section.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>