[nfc] Change all FIRRTL indentation to 2-space
Change all FIRRTL code examples to use 2-space indentation.  Previously
this was using a mix of 3-space (predominantly) and 2-space.  This does
not change the indentation used in the "Details about Syntax" section
which is intentionally written to highlight how indentation works and
can legally vary from block to block.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>