[major] Remove Fixed Point Types (#31)
Remove Fixed Point and any references to it from the FIRRTL
specification.  This (and intervals) was not a FIRRTL feature that
achieved significant traction nor was it ever load-bearing in the same
way that the rest of the specification has been for many years.

This is not in any way intended to be the end of Fixed Point
representations in FIRRTL (or Chisel)!  There are planned future
improvements to provide Fixed Point libraries either directly in Chisel
or via user-defined FIRRTL types once we sort through what that should
look like.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>