[major] Change connect to be truncating
Change the FIRRTL spec to state that connects are truncating (like
partial connects).  This has always been the way that connects work in
the Scala-based FIRRTL Compiler and the spec was always incorrect here.

My plan here is to remove partial connect and eventually add a "strict
connect" operator that does neither truncation nor sign
extension.  (This has been used for a while internal to the MLIR-based
FIRRTL Compiler to great benefit.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>