[Major] Explicit compile time constant expressions. (#67)
Add a constant type modifier for compile-time evaluation (under whole-world assumptions). Constant types in external ports implies that uses of constant types are not guaranteed to be know at compile time (this is not a mechanism to guarantee constant propagation).

Asynchronous resets are updated to take constant reset values. This moves ad-hoc checks which exist in implementations now to type checks. This is not sufficient for some tools, but is a strict improvement on the current implementations.

The change of asyncreset value is what makes this a major change.