[patch] Clarify Int/String Types, clarify SE-Ints
Add language that explicitly defines literal types:

  1. Integer literals
  2. String literals
  3. Raw String literals
  4. String-encoded integer literals

This only clarifies existing behavior which was left ambiguous or
documented in different locations in the spec.

Define that string-encoded integer literals can be used any place that
an integer literal can except where specified otherwise.  This is used
to differentiate where an external module parameter that looks like a
string-encoded integer literal is still a string.  This is also used to
preserve the existing SFC behavior where a string-encoded integer
literal is not allowed to be used for memory integers.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>