[major] Switch to 3.0.0 Syntaxes
Remove the old connect ("<="), invalidation ("is invalid"), and register
with reset ("with: (reset => (resetSignal, resetValue))") syntaxes in
favor of syntax which does not require lookahead when parsing (connect and
invalidate) and simpler syntax (register with reset).  The new syntaxes,
which wer added as "alternative syntaxes" previously, are "connect",
"invalidate", and "regreset".

This is a major breaking change that becomes active in the 3.0.0 release.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>