[minor] Add "regreset" alternative syntax
Add alternative syntax for specifying a register with a reset.  This is
done to simplify parsing, readability, and writability.  This also brings
the register more in line with other FIRRTL operations.

The new syntax looks like:

    regreset a: UInt<8>, clk, rst, init

The old syntax that this is an alternative to is:

    reg a: UInt<8>, clk with: (reset => (rst, init))

This syntax is intended to be introduced in the v2.3.0 minor release and
made mandatory in the v3.0.0 release.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>