[major] Require Bundle fields are legal IDs
Change the spec to require that Bundle fields are legal identifiers.
Currently, the SFC has historically accepted anything, including numbers
for fields.  This stems from the MixedVec Chisel utility which emits
numeric fields.  Change this to make lexing FIRRTL easier.  Currently,
there is extreme ambiguity for something like "0.0".  This could be either
a subfield or a floating point number.  By restricting subfields to
require legal identifiers, this becomes illegal and the ambiguity is
resolved.

Note: the EBNF description is already correct here.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>