[major] Add literal identifiers, e.g., `0`
Add literal identifiers which are identifiers that are escaped with
backticks.  This is used to encode an identifier which (extremely
narrowly) is an identifier but starts with a leading number.  This enables
extremely simple lexing and avoids any confusion with other possible
number types, e.g., floats or versions.

Add language that describes how this works for subfields.  E.g., to
subfield into an aggregate "foo" with field "`0`" you would do "foo.`0`".

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>