[nfc] Change LaTeX Template with spec/abi versions
Change the format of the revision-history.yaml file to put version
information under "spec" or "abi" keys.  Update the LaTeX template to
take a key ("revisionHistorySpec" or "revisionHistoryAbi") that will only
include version information from the "spec" or "abi" keys.  This enables a
unified revision-history.yaml for co-versioning of different documents.

This assumes that we want to co-version everything, e.g., the
specification and the ABI, but split them into separate documents.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>