[patch] Simplify Last Connect, rm Statement Groups
Simplify the explanation of "last connect semantics".  Remove the
explanatory feature "statement groups".  This is only used to explain last
connect semantics, but is unnecessary and is not used in the similar
explanation of "conditional last connect semantics".  Move the "last
connect semantics" section up one level to be a sub-section of the
"connect" section.

This frees up language in the spec to talk about "groups" without
conflicting with the removed "statement groups".

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>