[minor] add enumeration types and match statements (#111)
Add product types to firrtl. These are called enums after rust and swift. Other languages might call them discriminated unions. The type carries a discriminator which is a set of identifiers, each a tag corresponding to one variant. The type also carries a union, with each variant of the union corresponding to a specific tag. A value of enum type will have one tag and the data associated with that tag. Match statements provide a way to decompose an enum. Enum construction is provided in the same manner as ground types.

An enum with empty fields (just tags) is expected to map to a Verilog enum. An enum with fields maps to a verilog tagged union in spirit, but due to almost no tools support synthesis of packed tagged unions (or even just packed tagged unions as a recognized type), there will be a more complex lowering which will be detailed in the ABI document.

Several things are intentionally left off at this point. These include controlling the tags' integer values, extracting the tag as an integer, casting enums to/from integers unsafely, and match failures (tag outside declared set of values).


---------

Co-authored-by: Andrew Young <youngar17@gmail.com>
Co-authored-by: Will Dietz <will.dietz@sifive.com>