[patch] Tweak mux selector width-inference behavior.
Don't require inferring selector to 1-bit:
* Support inferring to zero-bit (will be padded)
* Don't put constraints on operands ("back-propagation").
* Align with behavior in similar constructs: when, subaccess.

As-is selector can be entirely unconstrained otherwise,
but will inferred to one-bit when used in a mux.

This masks design bugs and has no known utility.

Further, by requiring inference to 1-bit this can cause hardware that otherwise
would be zero-width (and often elided) to become 1-bit and kept around.

Instead treat the selector like any other operand,
but still reject if selector is inferred to be larger than 1 bit.

See https://github.com/llvm/circt/issues/5444 for further discussion and
examples demonstrating the issue on current and historical implementations.