[major] Add public modules
Add a "public" modules to FIRRTL.  A public module is a module that is a
root of the circuit.  The notion of a circuit is then expanded to allow
for multiple roots and these roots may have no common children.  In
effect, a circuit is a collection of multi-rooted instantiation trees.

Modify the ABI to require that all public modules produce a filelist whose
name is derived from the public module's name.

Remove circuit name entirely as this has no bearing on the ABI anymore.
As a result, Annotations can no longer target the circuit.  Annotations
which did this previously now must be "no target annotations".  This was
always supported and was how Chisel has always emitted them.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>