[nfc] Make LaTeX warnings errors
This is done to catch things like undefined references which we don't want
leaking into the spec.  The output on an error isn't great.  You get the
useful latexmk output (e.g., indicating exactly what references are
undefined) followed by the entire output of the LaTeX build.  Developers
need to know to scroll to the top.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>