[nfc] Move contributors to JSON file
Instead of putting contributors as inline markdown, put them in a new JSON
file (include/contributors.json) and convert this to an itemized list in
the LaTeX template.  The template provides a new command `\contributors`
that can be used to get this list of contributors.

This is done to make this easier to maintain such that a CI script can
automatically update this.  JSON is easier to update than inline markdown.

Work towards #153.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>