[major] Rename "groups" to "layers"
Rename optional groups to layers.  This much better reflects what the
language feature actually is.  This also makes the focus on the
"layer" (previously "groupdecl") instead of the "layer block" (previously
"group").  This inversion fixes issues where the real concept is the
collection of optional functionality and not the blocks of optional
functionality that may or may not be in different modules.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>