[major] Finalize FIRRTL 4.0.0 Public Modules
Remove some of the changes for public modules to make these easier to
implement without dramatic changes to how Chisel is invoked and requiring
roll out of a lot of changes in FIRRTL compilers.  Specifically, continue
to keep the "main module" and thereby require that a FIRRTL circuit has a
single entry point.

The FIRRTL ABI is mostly unmodified with these changes.

The intent is to remove the circuit completely in FIRRTL 5.0 and delay
changes to Chisel and FIRRTL compilers until then.  (There's a lot in
FIRRTL 4.0 already and it would be good to cut it somewhere.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>