[patch] Break layer block text out into section
Move the text, with minimal modifications, that describes layer blocks out
of the layers subsection and into its own top-level section.  This is done
to keep the layers description tighter---layers declare the existence of
optional functionality while layer blocks define the optional
functionality.

This is done in preparation for broader updates to layers and layer blocks
that better define their semantics.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>