[nfc] Add ability to test code snippets
This demonstrates a flow that can be used to validate that all code
snippets used in the FIRRTL specification correctly parse.  This is
implemented using a simple pandoc filter to extract the bodies of all code
blocks with known languages (FIRRTL and Verilog), write them to the build
directory, and then run them all through an appropriate tool based on file
extension (firtool or verilator).

This may be a little cumbersome as every code example has to be completely
legal.  To alleviate this, this PR includes the ability to add special
delimiters, indicated by lines with the text `snippetbegin`/`snippetend` which
will cause the filter to extract only the lines between the delimiters.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>