[nfc] Add basic TextMate Grammar
Add a basic TextMate-compatible grammar for FIRRTL.  This is _not_
perfect.  It should get most things right.  However, a number of quirks in
the FIRRTL language make the TextMate format difficult to work with.

This is suitable as a building block of a VSCode extension.

This is specifically working towards enabling #54 which requires a
TextMate-compatible grammar.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>