[patch] Clarify existing when behavior
Add language that defines what conditionals are (they are a region where
the operations in the region change their behavior based on the condition)
and how different operations are affected by the condition.  This fills a
hole where the behavior of commands was not explicitly specified.  Add an
aside that clarifies that this means that a _trivial_ inlining will not
work as a user might expect.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>