[major] Adjust syntax of formal unit tests
Adjust the syntax of the `formal` declaration to no longer expect a
single fixed `bound = <N>` parameter, but instead accept a list of
user-defined parameters similar to an `extmodule`. The parameters can be
integers, strings, arrays, and dictionaries. They are passed through to
the tooling running the formal test. As we build out that tooling, we
may standardize some of the parameters.

Marking this as "major" since it technically is a breaking change, but
in practice there are no users of `formal` as of today and the original
syntax has never been released as part of the spec.