[major] Add special substitutions (#288)
Add special substitutions.  These look like `{{Name}}`.  Add two special
substitutions that we want to enable going forward:

  - `{{SimulationTime}}`
  - `{{HierarchicalModuleName}}`

This includes an escape sequence that can be used to get a literal `{{` in
the output.  Over time, these are expected to be changed to include more
special substitutions on an as-need basis.

This is a major change as it is backwards-incompatible with old FIRRTL
versions.  This will break compilation of any code which used `{{`
previously.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>