[minor] Add simulation unit test construct
Add a `simulation` unit test declaration analogous to the existing
`formal` construct. This allows simulation test harness modules to be
marked as entry points for simulation, which downstream tools can detect
and run. This is more or less a copy of `formal`, with a few added
constraints on the target module's ports and the exact simulation
schedule.