[major] Drop Circuit from Annotations
Drop mentions of the "Circuit" from annotations.  We want to move away
from this as the circuit name doesn't add anythiing meaningful other than
locking a given annotation file to a circuit.  However, given that we have
had the inline annotation format for a long time, this locking is
unnecessary.

Pragmatically, this makes it easier to generate annotations from Chisel
because the module name is always known, but the circuit name may not be.
E.g., when compiling in a D/I environment where definitions are built and
the instantiated in a separate circuit.

Note: the diff here is minimal.  The spec already has "targets" specified
_without_ a circuit name in their EBNF grammar description.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>