[major] Drop module name from inline layer macro
This PR changes the FIRRTL ABI for inline layers so that there is a single
macro which, when defined, will enable the inline layer across all compilation
units (circuits) in a design. This is done by removing the module-name from the
macro used to enable the layer.

The type system for FIRRTL ensures that a layer-guarded hardware value is only
accessed when the layer is enabled, but this is only true when a layer is
enabled or disabled universally.

By including the module name in the inline layer macro, we have a distinct
macro for enabling a layer for each module in a design, which would allow us to
enable a layer on a module-by-module basis, which is not desirable.

Additionally, the ABI for inline macro names now uses `$` as a deliminator for
name parts uniformly.

Before this PR, the macro format was:

```
layer_Module$Layer1$Layer2
```

Now, the format is:

```
layer$Layer1$Layer2
```

In a later change, we will be updating the ABI for ref names to use `$` as
well.