Prefer tee -a over >> so format diff is discoverable.
This has burned folks multiple times.

Posting it more prominently would be better, but at least
this way the diff isn't buried in an unreachable-from-PR location,
and will be visible when clicking through the failed job.

This helps occasional and frequent contributors alike.

Limit `tee -a` to just the `git diff`.

Per reviewer suggestion, thanks @seldridge.